Changes to make deployment more flexible and resilient
* Use of COMPOSE_PROJECT_NAME in env-gigadb instead of container_name in docker-compose:

I need to deploy multiple instances of the same stack of containers on the same machine.
The use of container_name precludes multiple instances as they would be a name clash.
Similarly the use of container_name prevents using docker-compose's scaling feature.
Finally, most of container_name values were identical to the service name.
Normally docker-compose creates name dynamically using the enclosing directory name for prefix
But I needed more control so the COMPOSE_PROJECT_NAME is a special env variable recognised
by docker-compose that sets the prefix of instanciated containers to its value.

* the sql files to bootstrap the gigadb db container are sourced from the gigadb website git checkout:

Because the database schema can change like normal code, there's a risk for the sql files in yii2-laradock to go stale.
It's also create a duplication. So there is now a SQL_DIRECTORY variable in env-gigadb
set by default to a gigadb-website checkout's sql directory

* Not hardcoding the path to gigadb-website in generate_config.sh

Instead, use the value of APPLICATION already in env-gigadb. This way, I can have two checkouts (different branches or different commmit version)
and instantiate the docker stack for each. of the checkout. so all instances of /vagrant have been replace by $APPLICATION in generate_config.sh and docker-compose.yml

* In generate_config.sh, the ftp examples files are not downloaded if they already present locally

* In generate_config.sh, instead of git cloning Yii 1, I download the release tarball

That way the script can be re-run many times without crashing at an attempt to clone in an already cloned directory.
Also the download won't if the release tarball is already present locally.

* Made sites-available read-only in the volumes config for nginx service in docker-compose.yml

If I need to change the nginx config in-container (for example for debugging/profiling),
any change will be made to the nginx site config files in yii2-laradock which is not desirable
as it will affect any future instances of the docker stack.
Instead,  to cater for in-container changes, I've added another read/writable volume linked to the DATA_SAVE_PATH.

* the vsftpd service was referencing hard-coded path for Yii dependencies in its volumes config

I've replaced them with YII_PATH

* In generate_config.sh, changed each line of sed based replacement to not execute sed at all
if the cp command failed.