nginx config with gigadb and yii2-laradock
We want the nginx site configuration in the provisioning system to not be altered whatever we do in the instanciated containers.
At the same time, we want the instanciation into a working web site to be fully automated,
once the configuration has been done.

Original state of the code would work just fine if the settings in .nginx/sites/gigadb.conf is enough for our use case.

If not (e.g: because we want to enable profiling, make the web site visible publicly under a FQDN or restrict a publicly facing web site),
we will have to edit the nginx config and that would alter the configuration from the provisioning system.

The current state of the code is an improvement, as the configuration mounted into the container
is read only and it still work fine for default use case.
But if we to change nginx behaviour we need to edit both nginx.conf and copy new site config in a new directory before editing to suit our needs.

This commit improves on that:
* default use case still work out-of-the-box (Gigadb website avilable on 192.168.42.10)
* the actual instance site configuration is saved under DATA_SAVE_PATH, which is a good practice already used in the other containerised apps of yii2-laradock
* the server_name is sourced from the HOME_URL variable already present with the appropriate default in env-gigadb
* If the only deviation from default use case is a different server name, the new instance will work out-of-the-box if HOME_URL has been configured accordingly
* if deviations from default use case are more significant, it's safe to edit the files in DATA_SAVE_PATH as they do not alter the configuration files from the provisioning system
* there is no need to do anything to nginx.conf