Further changes to make it work natively on macOS when using official Docker for macOS
in generate_config.sh, the way sed was used for pattern matching and replacement doesnt work on macOS
because macOS use the BSD sed whose -i option behave differently. If changing it to macOS idiomatic syntax
it breaks on Linux. The new syntax doesn't use -i and work with both BSD sed and Linux sed.
Also created SOURCE and TARGET in the gigadb sed-based config generation block to avoid repeating and for easy reading

Refactored the download of the Yii 1.1 framework so we can specify the version to download using the ENV variable YII_VERSION
The download and install of the yii 1.1 framework is now more robust. It is installed inside DATA_SAVE_PATH like any persistent container data of the container stack.
The in-container default path (YII_PATH) is by default non-specific to which version of yii 1.1 is installed (moved from /opt/1.1.16 to /opt/yii-1.1).
This is because GigaDB use yii 1.1.16 on production currently but the latest available 1.1.20 and there is 1.1.21 under developement, only 1.1.17+ supports PHP7+

set up default for the new ENV variables introduced with the code audit work, so the generate_config.sh still work when they are missing (for example if they were based on the original env-gigadb)

removed the SQL_DIRECTORY ENV variable I've introduced earlier as it is always going to be $APPLICATIONS/sql

known issue: I don't yet what to do with the call to yiic lesscompiler as it erroneously relied on the in-container yii path to be the same as the host path to the yii framework, so it's currently temporarily commented out until I fix it.