Compiling less file in Gigadb website
the problems with the original line

/vagrant/protected/yiic lesscompiler

which became

${APPLICATION}/protected/yiic lesscompiler

in my initial changes as part of the GigaDB code audit

firstly, is that it erroneously expect the in-container path to Yii be the same as on the host (/opt/yii-1.1.16)
on the host, we  should be free to have anywhere convenient but for sake of consistent practice with the rest of the stack,
the anywhere is wherever DATA_SAVE_PATH is.

secondly, it expects the host to have php installed

Even if i decide to use the official lessc binary, that expects the host to have NodeJS installed.

The elegant solution is to add one-off initialising service in docker-compose
from the official php-cli base image that will mount volumes from the app one-off service and  mount the Yii path
and then execute the Yii console command lesscompiler.