Enabled profiling of PHP web application with xhprof and xhgui
Used edyan/xhgui for visualisation.
Created a new dockerfile for php-fpm based on laradock's base image
(so it has already support for databases, gd, mcrypt and dev dependencies
and installed xhprof and mongo client in there.
Created a new compose file to compose a new stack of containers where profiling is enabled